Fix run-flexbot.sh

+++ run-flexbot.sh
@@ -1,7 +1,9 @@
-#!/bin/bash
+#!/bin/bash -x
 
 if [[ $# -ne 2 ]]; then
-    print "usage: $0 <ENV> <start|stop>\nwhere <ENV> is one of dev, prod"
+    echo "usage: $0 <start|stop> <ENV>"
+    echo "       where <ENV> is one of dev, prod"
+    exit 1
 fi
 
 IMAGE_NAME=yucht/flexbot
@@ -23,22 +25,25 @@ case $1 in
     ;;
   *)
     print "$1 is not one of dev, prod"
-    exit 1;
+    exit 2;
 esac
 
 IMAGE_TO_PULL="$IMAGE_NAME:$TAG"
+TIMEZONE_CONFIG="-v /etc/localtime:/etc/localtime -v /etc/timezone:/etc/timezone"
 
 case $2 in
   start)
     docker pull $IMAGE_TO_PULL
-    docker run -v $CONFIG_FILE:/flexbot/configuration/config.yaml -v $LOGGING_CONFIG_FILE:/flexbot/configuration/logging.yaml -v $LOG_FILE:/databricks/logs/slackbot.log -p $PORT:$PORT -d --name=$CONTAINER_NAME -t $IMAGE_TO_PULL
+    OLD_CONTAINER=$(docker ps -a | grep $CONTAINER_NAME)
+    if [[ -n "$OLD_CONTAINER" ]]; then
+        docker rm -f $CONTAINER_NAME
+    fi
+    docker run -v $CONFIG_FILE:/flexbot/configuration/config.yaml -v $LOGGING_CONFIG_FILE:/flexbot/configuration/logging.yaml -v $LOG_FILE:/flexbot/logs/slackbot.log $TIMEZONE_CONFIG -p $PORT:$PORT -d --name=$CONTAINER_NAME -t $IMAGE_TO_PULL
     ;;
   stop)
     docker stop $CONTAINER_NAME
-    sleep 5
-    docker rm -f $CONTAINER_NAME
     ;;
   *)
     print "$2 must be one of start, stop"
-    exit 2
+    exit 3
 esac
